[mojo-stdlib] Remove unconstrained mixed dtype pow overloads. (#38385)
This removes the `x**y` operations that don't constrain the RHS
type (but keeps the one with an Int RHS).  The problem with these
is that they are significant footguns in cases like `x**0.5`
because the compiler infers the RHS to be F64 (the materialization
type of FloatLiteral) regardless of what the LHS type is.

The pow() global function documents that the LHS and RHS types
must be the same, so use the compiler to enforce this.  I discovered
this while working on improving type inference.

QUANTIZATION_ORIG_REV_ID: 30db26e3fe29a6818cbe6df64ccb0c6f329d0522

+++ mojo/kernels/test/quantization/test_per_channel_grouped_4bit.mojo
@@ -39,8 +39,8 @@ fn _run_test_quant[group_size: Int, tolerance: FloatLiteral]() -> Bool:
         var l2_norm_err = (
             (input_vec - decoded_result) * (input_vec - decoded_result)
         ).reduce_add()
-        l2_norm_err = l2_norm_err ** Float32(0.5)
-        var l2_norm_input = (input_vec * input_vec).reduce_add() ** Float32(0.5)
+        l2_norm_err = l2_norm_err**0.5
+        var l2_norm_input = (input_vec * input_vec).reduce_add() ** 0.5
         var rel_l2_norm = l2_norm_err / l2_norm_input
         print("rel-l2-norm      :", rel_l2_norm)
